auth_service/requirements.txt
Flask
requests
python-dotenv
Flask-SQLAlchemy
PyJWT
gunicorn
